correcting some feed bugs
