+++ public/app/mage/feed/observation-feed.component.html
@@ -49,6 +49,6 @@
     <div event="$ctrl.event" observation-news-item="observation" selectedobservation-form="$ctrl.event.form" container-element="side-bar"></div>
   </div>
 </div>
-<button class="mdc-fab app-fab--absolute" aria-label="New Observation" ng-click="$ctrl.createNewObservation()">
+<button class="mdc-fab app-fab--absolute" aria-label="New Observation" ng-click="$ctrl.onCreateNewObservation()">
   <span class="mdc-fab__icon material-icons record">add</span>
 </button>
